Flesh out tests some more, ensure # of concurrent clients & requests
Test for
- many requests on pre-established connections
- many requests on clients that connect at different times
- many requests which don't reuse a connection
